best structuring of AI logic
