+++ .gitignore
@@ -2,4 +2,4 @@ pkg/
 *.ods
 *.ods#
 src/github.com/
-quarto
+./quarto
